+++ sdk/resource/export_os_release_darwin_test.go
@@ -15,19 +15,6 @@
 package resource // import "go.opentelemetry.io/otel/sdk/resource"
 
 var (
-	SetDefaultOSProviders           = setDefaultOSProviders
-	SetOSProviders                  = setOSProviders
-	SetDefaultRuntimeProviders      = setDefaultRuntimeProviders
-	SetRuntimeProviders             = setRuntimeProviders
-	SetDefaultUserProviders         = setDefaultUserProviders
-	SetUserProviders                = setUserProviders
-	SetDefaultOSDescriptionProvider = setDefaultOSDescriptionProvider
-	SetOSDescriptionProvider        = setOSDescriptionProvider
-)
-
-var (
-	CommandArgs = commandArgs
-	RuntimeName = runtimeName
-	RuntimeOS   = runtimeOS
-	RuntimeArch = runtimeArch
+	ParsePlistFile = parsePlistFile
+	BuildOSRelease = buildOSRelease
 )