+++ sdk/resource/export_windows_test.go
@@ -15,19 +15,12 @@
 package resource // import "go.opentelemetry.io/otel/sdk/resource"
 
 var (
-	SetDefaultOSProviders           = setDefaultOSProviders
-	SetOSProviders                  = setOSProviders
-	SetDefaultRuntimeProviders      = setDefaultRuntimeProviders
-	SetRuntimeProviders             = setRuntimeProviders
-	SetDefaultUserProviders         = setDefaultUserProviders
-	SetUserProviders                = setUserProviders
-	SetDefaultOSDescriptionProvider = setDefaultOSDescriptionProvider
-	SetOSDescriptionProvider        = setOSDescriptionProvider
-)
-
-var (
-	CommandArgs = commandArgs
-	RuntimeName = runtimeName
-	RuntimeOS   = runtimeOS
-	RuntimeArch = runtimeArch
+	PlatformOSDescription         = platformOSDescription
+	ReadProductName               = readProductName
+	ReadDisplayVersion            = readDisplayVersion
+	ReadReleaseID                 = readReleaseID
+	ReadCurrentMajorVersionNumber = readCurrentMajorVersionNumber
+	ReadCurrentMinorVersionNumber = readCurrentMinorVersionNumber
+	ReadCurrentBuildNumber        = readCurrentBuildNumber
+	ReadUBR                       = readUBR
 )