+++ sdk/resource/os_test.go
@@ -30,10 +30,15 @@ func mockRuntimeProviders() {
 		func() string { return "LINUX" },
 		fakeRuntimeArchProvider,
 	)
+
+	resource.SetOSDescriptionProvider(
+		func() (string, error) { return "Test", nil },
+	)
 }
 
 func TestWithOSType(t *testing.T) {
 	mockRuntimeProviders()
+	t.Cleanup(restoreAttributesProviders)
 
 	ctx := context.Background()
 
@@ -45,6 +50,37 @@ func TestWithOSType(t *testing.T) {
 	require.EqualValues(t, map[string]string{
 		"os.type": "linux",
 	}, toMap(res))
+}
 
-	restoreProcessAttributesProviders()
+func TestWithOSDescription(t *testing.T) {
+	mockRuntimeProviders()
+	t.Cleanup(restoreAttributesProviders)
+
+	ctx := context.Background()
+
+	res, err := resource.New(ctx,
+		resource.WithOSDescription(),
+	)
+
+	require.NoError(t, err)
+	require.EqualValues(t, map[string]string{
+		"os.description": "Test",
+	}, toMap(res))
+}
+
+func TestWithOS(t *testing.T) {
+	mockRuntimeProviders()
+	t.Cleanup(restoreAttributesProviders)
+
+	ctx := context.Background()
+
+	res, err := resource.New(ctx,
+		resource.WithOS(),
+	)
+
+	require.NoError(t, err)
+	require.EqualValues(t, map[string]string{
+		"os.type":        "linux",
+		"os.description": "Test",
+	}, toMap(res))
 }